Refactoring to change to base url, adding health check for API.
