+++ api/src/main/java/gov/cms/ab2d/api/security/SecurityConfig.java
@@ -17,11 +17,11 @@ import org.springframework.security.web.authentication.logout.LogoutFilter;
 
 import java.time.Duration;
 
-import static gov.cms.ab2d.api.util.Constants.FHIR_PREFIX;
-import static gov.cms.ab2d.api.util.Constants.ADMIN_PREFIX;
 import static gov.cms.ab2d.api.util.Constants.ADMIN_ROLE;
+import static gov.cms.ab2d.common.util.Constants.ADMIN_PREFIX;
+import static gov.cms.ab2d.common.util.Constants.API_PREFIX;
+import static gov.cms.ab2d.common.util.Constants.FHIR_PREFIX;
 import static gov.cms.ab2d.common.util.Constants.SPONSOR_ROLE;
-import static gov.cms.ab2d.api.util.Constants.API_PREFIX;
 
 
 @Configuration
