+++ api/src/main/java/gov/cms/ab2d/api/util/Constants.java
@@ -1,5 +1,8 @@
 package gov.cms.ab2d.api.util;
 
+import static gov.cms.ab2d.common.util.Constants.API_PREFIX;
+import static gov.cms.ab2d.common.util.Constants.FHIR_PREFIX;
+
 public final class Constants {
 
     public static final String GENERIC_FHIR_ERR_MSG = "The body will contain a FHIR " +
@@ -7,12 +10,6 @@ public final class Constants {
             "https://www.hl7.org/fhir/operationoutcome.html " +
             "Please refer to the body of the response for " +
             "details.";
-
-    public static final String API_PREFIX = "/api/v1";
-
-    public static final String FHIR_PREFIX = "/fhir";
-
-    public static final String ADMIN_PREFIX = "/admin";
 
     public static final String ADMIN_ROLE = "ADMIN";
