+++ api/src/test/java/gov/cms/ab2d/api/controller/RoleTests.java
@@ -22,8 +22,7 @@ import org.testcontainers.junit.jupiter.Testcontainers;
 import java.io.InputStream;
 import java.util.List;
 
-import static gov.cms.ab2d.api.util.Constants.*;
-import static gov.cms.ab2d.common.util.Constants.SPONSOR_ROLE;
+import static gov.cms.ab2d.common.util.Constants.*;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
