+++ common/src/main/java/gov/cms/ab2d/common/util/Constants.java
@@ -21,4 +21,10 @@ public final class Constants {
     public static final String ADMIN_ROLE = "ADMIN";
 
     public static final String EOB = "ExplanationOfBenefit";
+
+    public static final String API_PREFIX = "/api/v1";
+
+    public static final String FHIR_PREFIX = "/fhir";
+
+    public static final String ADMIN_PREFIX = "/admin";
 }
