+++ e2e-test/src/test/resources/prod-config.yml
@@ -1,3 +1,2 @@
 okta-url: 'https://test.idp.idm.cms.gov/oauth2/aus2r7y3gdaFMKBol297/v1/token'
-ab2d-api-url: 'http://ab2d-dev-204191066.us-east-1.elb.amazonaws.com/api/v1/fhir/'
-ab2d-admin-url: 'http://ab2d-dev-204191066.us-east-1.elb.amazonaws.com/api/v1/admin/'
+base-url: 'http://ab2d-dev-204191066.us-east-1.elb.amazonaws.com'
